chore(docs): New version script for release-please (#13503)
The currenet release-please version fetching script is not deterministic
as it grabs latest releases from the github api. There is a chance that
it gets versions that are not in the docs `versions.json`, which would
result in failing docs builds.

This new script updates the versions.json file deterministically,
passing the version number being released as part of the release in
which it's run.

As written in this PR, the docs should update to the new version as part
of a sandbox release. Since the alpha-testnet releases aren't automated,
those will still have to be manual docs updates.