feat(contracts): static + snapshotted validator set (#13046)
fixes: https://github.com/AztecProtocol/aztec-packages/issues/8761

This version stores a history of the validator set, without changing any
of the apis. Follow up will include commitments to the committee

This Pr adds a SnappshottedAddressLib that ensures the validator set
size and members cannot change during an epoch.

In its current form, calculations that rely on attesters.size can be
incorrect as it's size can change during the life time of an epoch. This
code snapshots the validator set such that queries to the attestor set
will remain fixed during the current epoch, and changes to the set will
only occur over the epoch boundary.