fix: only clear trailing zeros on logs (#13727)
We used to assume we'd only ever see zeros in a log due to padding, but
we now have scenarios were users put structs in a log that have some
zero fields, e.g. a bignum value with multiple limbs for which some of
them are zero when completing a partial note. If we knew the log length
as per #11636 this would be a non-issue, but for now we need to only
remove zeros _at the end_ of the array so that we can tell users that
their logs will be fine as long as it has a non-zero field at the end.