refactor: bb.js non-inlined web workers (#13736)
Inlined web workers are completely unsupported in browser extensions,
which created problems for external teams. This PR moves to a more
modern approach, leveraging the integrated support in latest versions of
`webpack` for web workers.

Unfortunately, I've run into quite a few problems implementing this
approach:

* Webpack didn't support proper module workers up until recently,
forcing me to update it (so it wouldn't transpile the imports to the
unsupported `importScripts`). Took the opportunity to equalize versions
accross the board.
* There's a bug in webpack that forbids loading them asynchronously and
threw me for a while: https://github.com/webpack/webpack/issues/17014.
Solution is to carefully handle our dynamic imports so they're only
applied to the wasm files.
* We were using `worker-loader` in `webpack` 5, which is explicitly
deprecated, but we needed it for the inlining. This gets rid of it.
* Unfortunately again, new webpack handles `.d.ts` and `.d.ts.map` files
differently now, which was causing trouble for downstream apps that also
used webpack. Test apps and boxes have been updated to ignore them at
build time (since they're only needed during development). Vite has no
issues.
* Took the opportunity to clean up our API, since we were never
initializing the `SyncApi`with a worker and conversely, the `Async` API
with a plain wasm module.