feat: hide Translator accumulated_result  (#13792)
`TranslatorCircuitBuilder` is responsible for computing the evaluation
at some challenge `x` of a batched polynomial derived from the `UltraOp`
version of the op_queue. This value gets sent to the
`TranslatorVerifier` as part of the proof and hence needs to not leak
information about the actual ops (explained in more detail as a comments
in the code). The PR resolves issue
https://github.com/AztecProtocol/barretenberg/issues/1368 and also
removes some left over ops that were just avoiding point at infinity
issues and are not necessary anymore.