feat: add a ultrahonk proof verification noir library (#14316)
When there's changes to the ultrahonk proof format, users are not
currently made aware of this until they try to verify a recursive proof
and bb blows up.

A nicer way to deal with this is for the crypto team to publish a small
wrapper around Noir's stdlib `verify_proof` function which sets all of
the proof and vk lengths. This means that if a user pulls the equivalent
version of this library to match the version of bb then they will get an
error much earlier if it's incompatible.

---------

Co-authored-by: AztecBot <tech@aztecprotocol.com>
Co-authored-by: ledwards2225 <l.edwards.d@gmail.com>