refactor!: helpers for  `AztecNode::getLogsByTags` and related oracle change (#14459)
Fixes #11639

Nowhere in our codebase we wanted to receive both public and private
logs and for this reason I separated the 2 APIs. This was especially
problematic in Noir (getLogByTag function) as public and private logs
have different lengths.

For the public logs we also wanted filtering by contract address so I
modified the API to do that.

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>