feat: completing partial notes in private (#14533)
Fixes #14362

Implements partial note completion in private. This PR is quite big but
I decided to just keep it like that because:
1. we currently don't have a way to unit test oracles,
2. testing partial notes flows in TXE test [doesn't really
work.](https://github.com/AztecProtocol/aztec-packages/issues/10289).

This prevented me from implementing the oracles, the partial note
completion and Orderbook changes in 3 PRs as I needed them in one to
verify it works.

## ~Update~
~Mr. Nico said this is too ugly (😭) and that I should do
https://github.com/AztecProtocol/aztec-packages/issues/10273 first~

The above is addressed now.