feat!: committing to events (#14634)
Fixes #14367

Commits to events by inserting the event commitment into the nullifier
tree. Upon private event log processing we enqueue the serialized event
along with its event commitment for validation. Then when the validation
is triggered we check that the commitment is in the nullifier tree and
if yes we store the event in the PXE db. Unlike for notes the event hash
function is fixed so we don't need to generate any additional code with
macros (like we do for notes with `_compute_note_hash_and_nullifier`
contract library function).

## Note on randomness in the payload
I decided to auto-inject the randomness into the payload instead of
having it be as a struct member in the event because:
1. If I included it in the struct it would break a ton of stuff (partial
notes, the bytecode emission related functionality etc.),
2. it would put the burden on the dev to include and if he didn't he
would shoot himself in the foot (privacy leak + reverting txs due to
potential nullifier collisions),
3. overall the API would be just uglier and the diff much larger.

Thanks to this ^ devs will not need to update their contracts.

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>