fix: Handle missed l1-to-l2 messages in archiver (#15116)
Fixes how the archiver handles a missed L1-to-L2 message caused by an L1
reorg. The archiver was updating the L1 syncpoint (ie last L1 block
checked) when there was no changes in the Inbox state, which means that
if a new message landed _before_ the syncpoint due to an L1 reorg, the
archiver would not see it.

This PR changes it so the syncpoint is only updated when messages are
actually retrieved. A downside for this change is that `inbox.getState`
gets called on L1 on every archiver loop, instead of once per L1 block,
but we can revisit when (if) we optimize the archiver's usage of L1 RPC.