feat: move authwits into aztec-nr (#15244)
As a stepping stone towards human readable authwits, I had to move the
library into `aztec-nr`. This is because the macro code that allows
exporting the ABI has to run at the contract level, and in general it
didn't make sense that `aztec-nr` knew about authwits with them being an
external library with a circular dependency.

Also moved the event selector out of protocol types and into aztec.nr,
since it's NOT a protocol concept anymore.

---------

Co-authored-by: Mitchell Tracy <mitchellftracy@gmail.com>
Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>