feat!: constrained GetEnvVar opcode in AVM (#14121)
Constrains the GetEnvVar opcode in the AVM and implements it in C++
simulation. It is implemented within the execution component. In other
words, it does not have a dedicated simulation component, and its PIL is
virtual to the execution trace.

This PR also adds `transaction_fee` and `global_variables` to context so
that simulation can access them anywhere. Lastly, this PR forces
teardown to have the same `transaction_fee` as the fee collection phase
(except when teardown is absent/padded).

Follow-up work:
1. Instead of using `should_resolve_address` and `sel_addressing_error`
here, execution should have something like `sel_dispatch_opcode`.
3. BUG: when this opcode is executed with enum `*GasLeft`, it should
return a `u32`
4. When this opcode is executed with enum `isStaticCall`, it should
return a `u1`
5. Rename `feePer*Gas` to `baseFeePer*Gas` to make room for (maybe
someday) `effectiveFeePer*Gas`.