chore: removed sendRequest (#15371)
Removes `sendRequest`, as per
[issue](https://github.com/AztecProtocol/aztec-packages/issues/15149#issue-3159433807):

`sendRequest` is problematic for the following reasons:

1. It sends a request to peer(s) randomly. This is suboptimal.
2. It tries N peers one by one until it times out. But in practice,
because of the default timeouts configured, it will have time to send
requests to only 2 peers at most.

I think the best course of action is to delete this method for the
following reasons:

1. The logic of which peer(s) we send requests to should be handled on a
per-request-type (topic) basis. This is proven/showcased by the fact
that we are not using this method at all for Status and Goodbye, but
reaching for the lower-level sendRequestToPeer. For fetching missing
transactions, this method is useless because, as mentioned previously,
we are only querying 2 random peers in practice.
2. This method is only called from requestTxByHash, which, as far as I
can trace the code, is not used (nothing is calling it).

Given points 1 and 2, I'd delete this.

### Note

When using `sendRequest` instead of `sendRequestToPeer`, we are losing
some of the niceties and have to deal with the buffers directly, for
example:

```TS
if (response && response.data.length > 0) {
          const object = subProtocolMap[subProtocol].response.fromBuffer(response.data);
          // The response validator handles peer punishment within
          const isValid = await responseValidator(request, object, peer);
          if (!isValid) {
            throw new InvalidResponseError();
          }
          return object;
        }
      }
    };
```

This is part of the code from `sendRequest`. I was thinking about adding
this code to either `sendRequestToPeer` or creating a new function
similar to `sendRequest` that wraps the `sendRequestToPeer` and deals
with this.

The problem is:
1. If I add it to `sendRequestToPeer`, I'll be breaking some of the
existing logic (around e.g. `STATUS` and `GOODBYE`)
2. If I create this new function, it will already be dead code, as it
won't be used outside of the tests.

I think the current solution is OK, and the caller of the
`sendRequestToPeer` should take care of this per their specific logic
(as `GOODBYE` and `STATUS` do)