fix: Proper slot start time in sequencer (#15471)
Fixes the slot start time computed on the sequencer.

Let's say we have the following, with genesis at L1 block 1, L1 block
time 10s, L2 block time 20s:

| Time    | L1 block | L2 slot | Notes |
| -------- | ------- | ------- | --- |
| 10  | 1 | 0 | |
| 20 | 2 | 0 | L2 block for slot 0 lands, start building for L2 slot 1
now (once we have sync'd) |
| 30 | 3 | 1 | Send the L1 tx for L2 slot 1 now (could have a few
seconds leeway) |
| 40  | 4 | 1 | The L1 tx for L2 slot 1 gets mined here |
| 50  | 5 | 2 |
| 60  | 6 | 2 |

The sequencer would compute the start time of the L2 slot as the L1
timestamp in which the first L1 block of the L2 slot is mined. Above,
the result of `getSlotStartTimestamp(1)` would be 30.

The sequencer timetable then works by checking how far into the slot
each state transition is, considering an L2 block time of 20s. In a
non-anvil chain, we assume L1 tx propagation requires a full L1 slot
(10s here), which means that the L1 tx for block proposal needs to be
sent at least 10s before the last L1 block of the L2 slot is mined. In
the example above, the L1 tx for slot 1 needs to be sent at ts=30.

However, if the sequencer computed the start of the slot as ts=30, the
timetable would be 10s off, believing it can send the L1 tx up until
ts=40. This PR should fix it, and set `getSlotStartTimestamp(1)` to
ts=20 instead of ts=30.