chore: validate note logs in reset (#15550)
Validating all note logs (`log.note_hash_counter != 0`) are linked to
existing note hashes in the reset circuit, so that we can check all the
logs at once when a reset is run. Previously it was checked in init and
inner, for all the logs emitted from a private function.

In doing so, we also allow logs to be linked to a note hash created in a
later function call.