feat!: commence call interface unification (#15894)
You can just do things. In this case, delete ~700 lines and end up with
a net improvement.

This deletes all of the `Void` variants of the call interfaces, in turn
deleting the `call_<>_void` variants, nicely unifying the testing
experience - we now just `call_public` or `call_private` without
worrying about return types. I also simplified `ReturnsHash` a bit since
it had three ways of achieving the same simple tasks, and added some
tests and clarifying comments about how the `()` case is handled.

There's lots more to improve on the `CallInterfaces`, and many more
removals to be done, but this initial pass is worth doing separately
because it directly improves `TestEnvironment`'s API.