feat(avm)!: memory multi permutations (#16586)
# Multi permutation builder (tracegen)

This component solves the following problem. Suppose you have `n`
permutations (P1, ..., Pn) such that
* The destination tuple columns (and therefore table) are the same for
all of them (e.g., mem.address, mem.value).
* There is a "global" destination table selector (mem.sel)
* The source selectors are (potentially) different, and **already set**
(e.g., execution.sel_addressing_base,
execution.sel_addressing_indirect_0..., sha256.mem_op_1...)
* The destination selectors are different, but UNSET. E.g.,
mem.sel_addressing_base, mem.sel_addressing_indirect_0...)

You want to find, for each source row of each permutation, a row in the
destination table and set the right destination selector. For the next
source row or permutation, you should not be able to consider the
"already used" destination rows, because this is a permutation.

## Solution

The multipermutation builder is templated over a list of permutation
settings. It builds an index of the destination table, and basically
applies the algorithm mentioned above.

## Example in this PR

For this PR, I made the required changes for Addressing, Keccak(mem) and
Sha256. Note, especially for sha, that we didn't need separate SOURCE
selectors because these are separate permutation instances.

## Downside

All of this gets currently done in one thread, and is also centralized
in the memory trace builder (as opposed to the sources). This might be
possible to change later.