feat!: verifying civc in tx base (#16713)
Remove the tube and verify `hiding_kernel_to_rollup` directly in private
tx base rollup.

Create a public tube circuit in noir that verifies
`hiding_kernel_to_public`. The public tube proof is verified in the
public tx base rollup.

Remove the fake tube vks and use the pre-generated vks in
`noir-protocol-circuit` for verification. All circuits now ensure that
the vks are in the vk tree.

Change `ClientIvcProof.ts` to contain proof fields instead of buffer.
The proof fields will be used by the orchestrator to construct the
private inputs to be recursively verified in private base or public
tube.

Co-authored-by: Leila Wang <leizciw@gmail.com>