feat: Allow vetoer to disable slasher temporarily (#16971)
Allows the vetoer to disable slashing altogether temporarily. This
allows the vetoer to stop all slashes if a bug is detected that
continuously triggers slashes that are not valid, so the vetoer does not
need to continuously vetoed every individual slash until the issue is
solved.

Slashing is disabled for `slashingDisableDuration` (immutable, set
during deployment, defaults to 5 days) and then automatically resumed.
Vetoer can extend the disabled period at any time, and can also resume
the slashing.