refactor: TestWallet API cleanup pt. 1 (#17283)
Followup work of
https://github.com/AztecProtocol/aztec-packages/pull/17228

In this PR I drop `getPXEInfo` and `getContracts` from `TestWallet` +
add random docs improvements.

`getPXEInfo` was just a tech debt as it was only used to check if node
has the same protocol contract addresses which is quite non-sensical
because those addresses are no literally 1, 2, 3, 4, 5... I assume this
check is there from time when they didn't have these hardcoded
addresses.

`getContracts` was used just in one test and the check there was not
really that important so I just dropped it.