feat: add docs reference frontmatter field (#17812)
The problem is that there are certain docs pages that are tightly
coupled with certain code reference files, although not exact lifts.
(i.e. interfaces, and node reference apis). If the code changes it's not
easy to track what areas in docs need to change as well.

@critesjosh suggested we added a reference field on docs frontmatter.

This field should support adding pages from the monorepo, and if these
pages change in a PR, to send a review to devrel to check the changes
and update the relevant tightly coupled docs.

This PR provides a basic implementation to do so. Users can add the
`reference` field on the frontmatter, as a list of relative paths to
source, and given changes in a PR, this will run as part of the docs
bootstrap and alert devrel to any changes.