chore: Try deflake inactivity slash tests (#18623)
We were slashing an online validator due to inactivity beacause it
failed to assemble the first block in the first slot of the epoch,
because we were warping to when the epoch was already started, not
before.

This should give the validator enough time to assemble the block and get
the proposal broadcasted.