feat: allow timestamp to be equal to the previous timestamp (#18939)
A block's timestamp can now be greater than or equal to the previous
block's timestamp.
Since the timestamp in the checkpoint header is the timestamp of its
last block. This enables using the same timestamp for all blocks in a
checkpoint, so that we don't have to decide whether a block will be the
last one before building it.