chore: domain separator tweaks (#19199)
Working towards improving domain separation.

(Getting ci to run the tests for me, to see what mistakes I've made)

In this PR:

- Remove unused domain separators from constants.nr
- Add some comments to constants.nr (and some physical reordering to
group them better... this is fine because the eventual goal is to use
strings for domain separators instead of brittle enums).
- Rename from `GENERATOR_INDEX__` to `DOM_SEP__`, because "generator
index" is carried over from when we used to use pedersen hashes and
makes no sense anymore.
    - Most of the line changes of this PR come from this.
- Renamed one of the separators from `__CONTRACT_LEAF` to
`__CONTRACT_CLASS_ID` to help my brain.
- Created `domain_separators.nr`, which contains dormant tests to
generate "proper" domain separators using strings. This PR won't go as
far as instating those strings, because all the hard-coded constants in
all tests in the repo will break. That'll be a fun task for someone in
January.

Future domain separation tasks for another PR:
- instate the domain separation strings.
- Claude says there are poseidon2 hashes which aren't using any domain
separator, which could be at risk of collisions. We'll need to add
those.