fix(docs): polishing rough edges (#19245)
On this PR we polish some rough edges:

- We unify the "CLI" references under the "CLI" on sidebar. Talked with
@critesjosh.
- We obviously modify auto-generation of references to work with the new
structure.
- We've moved the sign with FaceID into "Tutorials" (yikes, it needs a
better place)
- And we improve the installation of aztec's tooling by directly using
the version that is intended. This avoids having some problems than what
we have right now (the stable version of the install sh tries to install
`aztec-nargo`, but that install doesn't exist for the latest version, so
it breaks). This makes it so we point exactly to the install script we
need for the intended version.